refactor(tests): reorganize Azure OpenAI tests following Rust best practices
- Move unit tests to src/services/ai/azure_openai.rs (15 tests)
- Split large integration test file into 5 focused files:
  * azure_openai_api_integration_tests.rs (6 tests) - API interactions
  * azure_openai_error_handling_tests.rs (7 tests) - Error scenarios
  * azure_openai_response_parsing_tests.rs (3 tests) - Response parsing
  * azure_openai_deployment_tests.rs (6 tests) - Deployment scenarios
  * azure_openai_environment_tests.rs (1 test) - Environment variables
- Remove original monolithic azure_openai_integration_tests.rs file
- Preserve all 38 test cases with proper categorization
- Follow testing-guidelines.md principles for better maintainability
- Improve test organization and parallel execution support

Resolves #6
